# 🔐 Sprint 52: Platform Alignment & Audit Closure