--- a/configs/.env.example
+++ b/configs/.env.example
@@ -573,3 +573,27 @@ TELEMETRY_ENABLED=false
 # =============================================================================
 # END OF CONFIGURATION
 # =============================================================================
+
+
+# =============================================================================
+# STAGING ENVIRONMENT CONFIGURATION
+# =============================================================================
+# Example configuration for staging deployment (Railway, Render, etc.)
+# Set these in your platform's environment variables dashboard
+
+# Environment mode
+APP_ENV=staging
+
+# Telemetry configuration (Sprint 46-47: Prometheus + OpenTelemetry)
+TELEMETRY_ENABLED=true
+TELEMETRY_BACKEND=hybrid    # prom + otel (recommended for production-like environments)
+
+# OpenTelemetry tracing configuration
+OTEL_EXPORTER=console        # console (dev/CI) | otlp (staging/prod) | none
+OTEL_ENDPOINT=               # Set when using OTLP: http://tempo:4318 (HTTP) or http://tempo:4317 (gRPC)
+OTEL_SERVICE_NAME=djp-workflow-staging
+OTEL_TRACE_SAMPLE=0.05       # 5% sampling for staging (balance between cost and visibility)
+
+# Prometheus metrics configuration
+# Metrics available at /metrics endpoint (when TELEMETRY_ENABLED=true and backend=prom|hybrid)
+# Configure your Prometheus server to scrape: https://<your-staging-host>/metrics