--- a/requirements.in
+++ b/requirements.in
@@ -7,3 +7,4 @@ pandas>=2.0.0
 ujson>=5.0.0
 jsonschema>=4.0.0
 pydantic>=2.0.0
+httpx>=0.27.0  # Sprint 49 Phase B: Actions webhook adapter