--- a/requirements.in
+++ b/requirements.in
@@ -8,3 +8,10 @@ ujson>=5.0.0
 jsonschema>=4.0.0
 pydantic>=2.0.0
 httpx>=0.27.0  # Sprint 49 Phase B: Actions webhook adapter
+
+# Sprint 51 Phase 1: Database + Auth
+asyncpg>=0.30.0         # Async Postgres driver
+alembic>=1.13.0         # Database migrations
+sqlalchemy>=2.0.30      # ORM
+argon2-cffi>=23.1.0     # Secure password hashing
+psycopg2-binary>=2.9.9  # Postgres adapter for Alembic