--- a/requirements.in
+++ b/requirements.in
@@ -15,3 +15,6 @@ alembic>=1.13.0         # Database migrations
 sqlalchemy>=2.0.30      # ORM
 argon2-cffi>=23.1.0     # Secure password hashing
 psycopg2-binary>=2.9.9  # Postgres adapter for Alembic
+
+# Sprint 51 Phase 2: Rate limiting
+redis>=5.0.0            # Redis client for durable rate limiting