--- a/scripts/start-server.sh
+++ b/scripts/start-server.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+set -e
+PORT="${PORT:-8000}"
+exec python -m uvicorn src.webapi:app --host 0.0.0.0 --port "$PORT"