--- a/src/actions/adapters/__init__.py
+++ b/src/actions/adapters/__init__.py
@@ -1,0 +1,8 @@
+"""Action adapters for different providers.
+
+Sprint 49 Phase B: Independent adapter (Webhook only).
+"""
+
+from .independent import IndependentAdapter
+
+__all__ = ["IndependentAdapter"]