--- a/src/audit/__init__.py
+++ b/src/audit/__init__.py
@@ -1,0 +1,1 @@
+"""Audit logging for Sprint 51 Phase 1."""