--- a/src/auth/__init__.py
+++ b/src/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Authentication and authorization for Sprint 51 Phase 1."""