--- a/src/db/__init__.py
+++ b/src/db/__init__.py
@@ -1,0 +1,1 @@
+"""Database module for Sprint 51 Phase 1."""