--- a/src/limits/__init__.py
+++ b/src/limits/__init__.py
@@ -1,0 +1,1 @@
+"""Rate limiting for Relay actions API."""